django_mongo_db/employeeApp/api/serializers.py
from rest_framework import serializers

from employeeApp.models import Department, Employee


class DepartmentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Department
        fields = ('department_id', 'department_name')


class EmployeeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Employee
        fields = ('employee_id', 'employee_name', 'department', 'date_of_joining', 'photo_file_name')
