django_mongo_db/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
pymongo==3.12.3
pytz==2023.2
sqlparse==0.2.4
